backend/requirements.txt
flask
requests
flask_cors
google.generativeai
PyPDF2
streamlit
psycopg2